Refactor RangePrologue to a shared helper
